+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Master
+
+- Put all debug calls under the same namespace
+- Output config in logs
+
 # v0.2.1 (2018-12-22)
 
 - Implement proper logging via `debug` library (scope `@dgac/nmb2b-client`)